Move autonomous to separate folder
