+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autonomous/BlueLeft.java
@@ -1,0 +1,10 @@
+package org.firstinspires.ftc.teamcode.autonomous;
+
+import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+
+@Autonomous(name = "Red Left", group = "Autonomous")
+public class RedLeft extends AutoBase {
+    public RedLeft() {
+        super("red", "left");
+    }
+}
