+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autonomous/BlueRight.java
@@ -1,0 +1,10 @@
+package org.firstinspires.ftc.teamcode.autonomous;
+
+import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+
+@Autonomous(name = "Blue Right", group = "Autonomous")
+public class BlueRight extends AutoBase {
+    public BlueRight() {
+        super("blue", "right");
+    }
+}
