+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/autonomous/RedRight.java
@@ -1,0 +1,10 @@
+package org.firstinspires.ftc.teamcode.autonomous;
+
+import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+
+@Autonomous(name = "Red Right", group = "Autonomous")
+public class RedRight extends AutoBase {
+    public RedRight() {
+        super("red", "right");
+    }
+}
